Extracted verifier related helpers, type definitions and constants into separate folder.
