+++ core/App/contexts/configuration.tsx
@@ -1,12 +1,12 @@
 import { IndyPoolConfig } from '@aries-framework/core'
 import { createContext, ReducerAction, useContext } from 'react'
 
+import { ProofRequestTemplate } from '../../verifier/types/proof-reqeust-template'
 import { EmptyListProps } from '../components/misc/EmptyList'
 import { RecordProps } from '../components/record/Record'
 import OnboardingPages from '../screens/OnboardingPages'
 import { ScanProps } from '../screens/Scan'
 import { OCABundleResolver } from '../types/oca'
-import { ProofRequestTemplate } from '../types/proof-reqeust-template'
 import { PINSecurityParams } from '../types/security'
 import { SettingSection } from '../types/settings'
