+++ core/verifier/README.md
@@ -1,0 +1,4 @@
+## Verifier Module
+
+This folder contains set of helpers, type definitions, and constants allowing to add verifier capability into Aries Mobile Agent React Native.
+The folder is created to encourage the upcoming split of the project into sub packages.
